Reorganização dos handlers de listagem e catálogo
