+++ handler/create_category.go
@@ -1,0 +1,24 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/lcardelli/fornecedores/schemas"
+)
+
+// Cria uma nova categoria
+func CreateCategoryHandler(c *gin.Context) {
+	var category schemas.SupplierCategory
+	if err := c.ShouldBindJSON(&category); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if err := db.Create(&category).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao criar categoria"})
+		return
+	}
+
+	c.JSON(http.StatusCreated, category)
+}
