+++ handler/delete_supplier.go
