+++ handler/list_category.go
@@ -1,0 +1,19 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/lcardelli/fornecedores/schemas"
+)
+
+// Lista todas as categorias
+func ListCategoriesHandler(c *gin.Context) {
+	var categories []schemas.SupplierCategory
+	if err := db.Find(&categories).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar categorias"})
+		return
+	}
+
+	c.JSON(http.StatusOK, categories)
+}
