+++ handler/list_supplier.go
@@ -17,14 +17,18 @@ import (
 // @Success 200 {object} ListSuppliersResponse
 // @Failure 500 {object} ErrorResponse
 // @Router /suppliers [get]
+
+// Lista todos os fornecedores
 func ListSupplierHandler(ctx *gin.Context) {
 	var supplierLinks []schemas.SupplierLink
 
+	// Busca todos os fornecedores com a categoria e serviços correspondentes
 	if err := db.Preload("Category").Preload("Services").Preload("Services.Service").Find(&supplierLinks).Error; err != nil {
 		SendError(ctx, http.StatusInternalServerError, "Error listing supplier links")
 		return
 	}
 
+	// Cria uma slice de SupplierLinkResponse
 	var response []schemas.SupplierLinkResponse
 	for _, link := range supplierLinks {
 		externalSupplier, err := getFornecedorByCNPJ(link.CNPJ)
@@ -44,9 +48,11 @@ func ListSupplierHandler(ctx *gin.Context) {
 		})
 	}
 
+	// Envia a resposta de sucesso
 	SendSucces(ctx, "list-suppliers", response)
 }
 
+// Converte uma slice de SupplierService para uma slice de ServiceResponse
 func convertToServiceResponses(services []schemas.SupplierService) []schemas.ServiceResponse {
 	var responses []schemas.ServiceResponse
 	for _, service := range services {
