+++ handler/render_service_page.go
@@ -1,0 +1,41 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/lcardelli/fornecedores/schemas"
+)
+
+// Renderiza a página de cadastro de serviços
+func RenderServicePageHandler(c *gin.Context) {
+	userInterface, exists := c.Get("user")
+	if !exists {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "Usuário não autenticado"})
+		return
+	}
+	user, ok := userInterface.(schemas.User)
+	if !ok {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao obter informações do usuário"})
+		return
+	}
+
+	var categories []schemas.SupplierCategory
+	if err := db.Find(&categories).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar categorias"})
+		return
+	}
+
+	var services []schemas.Service
+	if err := db.Preload("Category").Find(&services).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao buscar serviços"})
+		return
+	}
+
+	c.HTML(http.StatusOK, "cadastro_servico.html", gin.H{
+		"user":       user,
+		"Categories": categories,
+		"Services":   services,
+		"activeMenu": "cadastro-servico",
+	})
+}
