+++ handler/update_category.go
@@ -1,0 +1,34 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/lcardelli/fornecedores/schemas"
+)
+
+// Atualiza uma categoria existente
+func UpdateCategoryHandler(c *gin.Context) {
+	id := c.Param("id")
+	var category schemas.SupplierCategory
+
+	// Busca a categoria no banco de dados
+	if err := db.First(&category, id).Error; err != nil {
+		c.JSON(http.StatusNotFound, gin.H{"error": "Categoria não encontrada"})
+		return
+	}
+
+	// Faz o bind do JSON recebido para a estrutura
+	if err := c.ShouldBindJSON(&category); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	// Salva as alterações no banco de dados
+	if err := db.Save(&category).Error; err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Erro ao atualizar categoria"})
+		return
+	}
+
+	c.JSON(http.StatusOK, category)
+}
